Update README.md
Find all the recurrent and transient classes.

Given that we start in state 2, what is the probability that we will reach state 0 before state 5?

What are all of the possible stationary distributions of this chain? Hint: Consider the recurrent classes.


(d) Suppose we start in the initial distribution 0 := 0 0 1 0 0 for some

2 [0; 1]. Does the distribution of the chain converge, and if so, to what?

2. Finite Random Walk

Assume 0 < p < 1. Find the stationary distribution. Hint: Let q = 1 p and =